my_blog/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
DateTime==5.5
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.1
fastapi-sessions==0.3.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pathlib==1.0.1
pluggy==1.5.0
pyasn1==0.6.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.38.2
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.12.2
uvicorn==0.30.6
zope.interface==7.0.1
